Initializes data structures for the simulation
Sets up the necessary data structures to manage the
philosophers, forks, and simulation parameters. This
includes initializing the table with the number of
philosophers, time to die, eat, sleep, and the optional
number of meals each philosopher must eat.

Also, initializes the mutexes for managing forks and
controlling access to shared resources, and properly links
philosophers to their forks.